Rename directory, from 'micorsoft' to 'microsoft'
