+++ packages/auth/src/microsoft/index.ts
