+++ packages/auth/src/microsoft/microsoft-auth-guard.ts
