+++ packages/auth/src/microsoft/microsoft.strategy.ts
